Using JSON from body, fixed and renamed endpoints
THe API now gets the JSON from the body of the HTTP request insted of URL encoding. This has lead to the removal of deserialization of JSON as the framework will do it for us. Endpoint names have been changed to account for the "FromBody" attribute added, which forces the endpoints to follow the standard practice of parameters